Changed routing of api
