+++ app/controllers/v1/messages_controller.rb
@@ -1,0 +1,14 @@
+module V1
+  class MessagesController < ApplicationController
+    def index
+      @message = random_message
+      render json: @message
+    end
+
+    private
+
+    def random_message
+      Message.order('RANDOM()').first
+    end
+  end
+end
