+++ config/routes.rb
@@ -5,10 +5,10 @@ Rails.application.routes.draw do
 
   root 'static#index'
 
-  namespace :api do
-    namespace :v1, defaults: { format: 'json'} do
-      resources :messages, only: [:index]
-    end
+ 
+  namespace :v1, defaults: { format: 'json'} do
+    resources :messages, only: [:index]
   end
+  
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
 end
